💄 style(refactor): migrate from chalk to ansi-colors
I've decided to switch to ansi-colors, as I felt that there is too many mix-match of colors and gradients, and seems way too obnoxious, even to me. This change does not affect any functionality, it simply improves the colors.

Sidenote: We still use Chalk for certain outputs, such as the badge code output, which is why we are keeping it around as a dependency.

---
Also included in this commit:

* ✏️ chore(errors): improve error messages
* 🐛 fix(formats): fix some formatting issues
---
* The formatting issue was a result of my (sorta failed) attempt at spacing out the outputs by adding either too many or one too ltitle blank console logs.

Signed-off-by: Inter <me@inttter.com>