💄 style(error): improve look of error messages
Thanks Consola!

Along with this:

* 🐛 fix(spinner): fix spinners not stopping after an error message

I forgot to actually...stop the spinners when an error was caught. This should stop the spinner now so it does not infinitely keep running even after an error.

* 👕 chore(lint): fix up some long-nested console log's

They looked very ugly and did not match the rest of the code. Might have missed one or two, but me scrolling through it right now, I do not see anything.