🧹 chore: use new logo
Last time the logo was updated was v3, and I made that one in a record time of about 30 seconds, so I figure it was time to make a (although similar in resemblence) new logo and switch the "color scheme" (?). Has already been applied to documentation and social preview.

Also made formatting/text adjustments to the README, since there was a lot of wasted space when it was centered.